+++ CFQ/Features/Sign/SignIn/ViewModel/SignInViewModel.swift
@@ -15,10 +15,12 @@ class SignInViewModel: ObservableObject {
     @Published var isConfirmScreenActive = false
     @Published var isEnabledResendCode: Bool = false
     @Published var numberTapResendCode: Int = 0
-
+    
     var timer: Timer? = nil
     var timerContString: Timer? = nil
     private var timerCancellable: AnyCancellable?
+    private var errorService = ErrorService()
+    
     @Published var timerString: String = "10"
     private var secondsElapsed = 10
     
@@ -71,7 +73,7 @@ class SignInViewModel: ObservableObject {
             formatPhoneNumber(for: phoneNumber), uiDelegate: nil
         ) { verificationID, error in
             if let error = error {
-                completion(false, self.errorAuthFirebase(error: error))
+                completion(false, self.errorService.errorAuthFirebase(error: error))
                 return
             }
             if let verificationID = verificationID {
@@ -98,7 +100,7 @@ class SignInViewModel: ObservableObject {
 
         Auth.auth().signIn(with: credential) { authResult, error in
             if let error = error {
-                completion(false, self.errorAuthFirebase(error: error))
+                completion(false, self.errorService.errorAuthFirebase(error: error))
                 return
             }
 
@@ -169,42 +171,4 @@ class SignInViewModel: ObservableObject {
         }
     }
 
-    private func errorAuthFirebase(error: Error) -> String {
-        let errorCode = (error as NSError).code
-        let errorMessage: String
-
-        switch errorCode {
-        case AuthErrorCode.invalidPhoneNumber.rawValue:
-            errorMessage = "Le numéro de téléphone n'est pas valide."
-        case AuthErrorCode.missingPhoneNumber.rawValue:
-            errorMessage = "Le numéro de téléphone est manquant."
-        case AuthErrorCode.quotaExceeded.rawValue:
-            errorMessage = "Quota dépassé pour l'envoi de messages."
-        case AuthErrorCode.tooManyRequests.rawValue:
-            errorMessage = "Trop de tentatives de connexion."
-        case AuthErrorCode.invalidVerificationCode.rawValue:
-            errorMessage = "Le code de vérification est invalide."
-        case AuthErrorCode.expiredActionCode.rawValue:
-            errorMessage = "Le code de vérification a expiré."
-        case AuthErrorCode.userDisabled.rawValue:
-            errorMessage = "L'utilisateur a été désactivé."
-        case AuthErrorCode.userNotFound.rawValue:
-            errorMessage = "Utilisateur non trouvé."
-        case AuthErrorCode.accountExistsWithDifferentCredential.rawValue:
-            errorMessage = "Un compte existe déjà avec un autre type de connexion."
-        case AuthErrorCode.internalError.rawValue:
-            errorMessage = "Erreur interne du serveur."
-        default:
-            if errorCode == 17020 {
-                print("Erreur réseau : \(error.localizedDescription)")
-                errorMessage = "Problème de connexion internet"
-            } else {
-                errorMessage = "Une erreur inconnue est survenue."
-            }
-        }
-
-        Logger.log(errorMessage, level: .error)
-
-        return errorMessage
-    }
 }
